pyproject.toml
[tool.poetry]
name = "airbnb-backend"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^9.4.0"
django = "^4.1.5"
djangorestframework = "^3.14.0"
black = "^22.12.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
